move old calc score function to design folder for archiving
