+++ design/calculate-score-v1/index.js
