+++ design/calculate-score-v1/readme.md
@@ -1,0 +1,4 @@
+# Calculate Score V1
+
+Obsolete and replaced with functions-app/calculate-score-V2
+preserved in case the average scoring is re-introduced
