+++ design/calculate-score-v1/v1.spec.js
