+++ functions/disable-functions.env
@@ -1,4 +1,3 @@
-AzureWebJobs.calculate-score.Disabled=true
 AzureWebJobs.census-import.Disabled=true
 AzureWebJobs.check-expiry.Disabled=true
 AzureWebJobs.check-started.Disabled=true
